fix(FeatureGoogleAnalytics): Refactor after review
